+++ README.md
@@ -1,11 +1,19 @@
 # react-training-project-1
+
 Training front-end skills in React
 
 ### useLayoutEffect Hook
+
 [reference](https://react.dev/reference/react/useLayoutEffect)
 
 ### useImperativeHandle Hook
+
 [reference](https://react.dev/reference/react/useImperativeHandle)
 
 ### hooks livecycle flow
+
 [diagram](https://github.com/donavon/hook-flow/blob/master/hook-flow.png)
+
+### Error Boundaries
+
+[reference](https://pt-br.react.dev/reference/react/Component#catching-rendering-errors-with-an-error-boundary)
